chore: clippy clean — 0 warnings across the workspace, decompiler output byte-identical
cargo clippy --workspace --all-targets: 143 -> 0 warnings. Behavior-neutral
throughout — suite 503/0 and the decompiler corpus (avg 0.9462, 57/60) are
byte-identical before and after.

- ~76 auto-fixed (needless_borrow, unnecessary_cast).
- 9 fixed in code (range_loop -> fill/enumerate in cfg.rs, a while-let, redundant
  struct-updates + a contains_key in tests) — all provably equivalent.
- 55 #[allow] with a one-line rationale at each site, NOT restructured, because
  the lint fights the faithful port: too_many_arguments / type_complexity mirror
  Ghidra signatures/types; if_same_then_else + nonminimal_bool (x6) are distinct
  cascaded conditions verified against Ghidra source (rangeutil.cc / ruleaction.cc
  / typeop.cc); dormant-but-ported dead code; one crate-level allow for the 34
  doc_lazy_continuation warnings (fire on prose module docs).

No ported function signature or faithful type was changed to appease a lint.

+++ crates/mosura/src/decompile/fspec.rs
@@ -273,6 +273,8 @@ impl ParamList {
             }
         }
         let pref = if float_count > int_count { type_class::FLOAT } else { type_class::GENERAL };
+        // faithful port of Ghidra's group scan; `i` is the group/slot index passed downstream
+        #[allow(clippy::needless_range_loop)]
         for i in 0..hitlist.len() {
             match hitlist[i] {
                 None => {
@@ -839,7 +841,7 @@ pub fn recover_input_params(f: &Funcdata) -> Vec<ProtoSlot> {
         if !vn.is_input() {
             continue;
         }
-        let size = vn.size as u32;
+        let size = vn.size;
         if !pl.possible_param(vn.loc, size) {
             continue;
         }
@@ -860,7 +862,7 @@ pub fn recover_output(f: &Funcdata) -> Option<ProtoSlot> {
         let o = f.op(op);
         if o.code() == OpCode::Return && o.num_inputs() > 1 {
             let v = o.input(1)?;
-            return Some(ProtoSlot { addr: f.vn(v).loc, size: f.vn(v).size as u32 });
+            return Some(ProtoSlot { addr: f.vn(v).loc, size: f.vn(v).size });
         }
     }
     None
